selection sort

+++ 2-selection_sort.c
@@ -1,27 +1,27 @@
 #include "sort.h"
 
 /**
-* selection_sort - sort array of integers with selection sort
-*
-* @array: array to sort
-* @size: size of the array
-*
-* Return: Nothing
-*/
+ * selection_sort - sort array of integers with selection sort
+ *
+ * @array: array to sort
+ * @size: size of the array
+ *
+ * Return: Nothing
+ */
 void selection_sort(int *array, size_t size)
 {
-  int i, j, min_index;
+	size_t i, j, min_index;
 
 	if (!array || !size || size < 2)
-    return;
-  for (i = 0; i < size - 1; i++)
-  {
-    min_index = i;
-    for (j = i + 1; j < size; j++)
-      if (array[j] < array[min_index])
-        min_index = j;
+		return;
+	for (i = 0; i < size - 1; i++)
+	{
+		min_index = i;
+		for (j = i + 1; j < size; j++)
+			if (array[j] < array[min_index])
+				min_index = j;
 
-    if (min_index != i)
-      swap(&array[i], &array[min_index]);
-  }
+		if (min_index != i)
+			swap(&array[i], &array[min_index]);
+	}
 }
